refactor: restructure authentication components
- Moved SuperTokensInitializer to utils for better organization.
- Introduced a new LoginPage component to replace the removed SignInPage and SignUpPage components.
- Deleted unused SignUpForm and SignUpView components to streamline the authentication flow.
