--- apps/web/src/app/auth/signin/page.tsx
@@ -1,6 +1,0 @@
-"use client";
-import LoginView from "@/features/auth/views/LoginView";
-
-export default function SignInPage() {  
-  return <LoginView />;
-}
