--- apps/web/src/app/auth/signup/page.tsx
@@ -1,5 +1,0 @@
-import SignUpView from "@/features/auth/views/SignUpView";
-
-export default function SignUpPage() {
-  return <SignUpView />;
-}
